% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{arma_irf}
\alias{arma_irf}
\title{ARMA Model Impulse Response Function}
\usage{
arma_irf(phi = numeric(), theta = numeric(), sigma = 1, h = 40)
}
\arguments{
\item{phi}{the AR coefficients in the difference equation notation.}

\item{theta}{the MA coefficients in the difference equation notation.}

\item{sigma}{the innovations standard deviation (see details).}

\item{h}{the forecast horizon. This may need to be tweaked to ensure that
the output is long enough to show the decay to zero. If only an MA model is
passed (phi is zero length), then h should be set to the MA order.}
}
\value{
A data.table with the horizon and responses, starting at period 0.
}
\description{
Calculates the impulse response function of an ARMA model.
}
\details{
The impulse response function (irf) of a time series model
measures the changes in the future responses of all variables in the system
when a variable is shocked by an impulse. For an ARMA model, this simply
transforms the ARMA part to an infinite MA representation. Finally, the responses
are scaled by the lower triangular factor in the Cholesky factorization of the
innovation variance (which is the standard deviation in this case).
For standardized shocks, where the variance is 1, the scaling is unity.
}
