% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{arma_modelspec}
\alias{arma_modelspec}
\title{ARMA Model Specification}
\usage{
arma_modelspec(
  y,
  order = c(0, 0),
  constant = TRUE,
  xreg = NULL,
  distribution = "norm",
  ...
)
}
\arguments{
\item{y}{an xts vector.}

\item{order}{the (p,q) ARMA order.}

\item{constant}{whether to estimate a constant (mean) for y,}

\item{xreg}{an optional xts matrix of regressors in the conditional mean
equation.}

\item{distribution}{a valid distribution from the available
re-parameterized distributions of the package.}

\item{...}{not used.}
}
\value{
An object of class \dQuote{tsarma.spec}.
}
\description{
Specifies an ARMA model prior to estimation.
}
\details{
The specification object holds the information and data which is
then passed to the maximum likelihood estimation routines.

The arma specification is meant for stationary data, not integrated. The constant
represents the mean of y so that the equation becomes:
\deqn{y[t] = mu[t] + phi[1](y[t-1] - mu[t-1]) + \dots + phi[p](y[t-p]-mu[t-p]) +
theta[1]e[t-1] + \dots + theta[q]e[t-q] + e[t]}
where \deqn{mu[t]} is the estimated mean of y plus any additional pre-lagged
regressors passed via the xreg argument. If the absence of regressors,
then \deqn{mu[t]} is constant across all time indices.
Initialization of the parameters is based on Hannan and Rissanen (1982).
}
\author{
Alexios Galanos
}
