% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{bread.tsarma.estimate}
\alias{bread.tsarma.estimate}
\alias{bread}
\title{Bread Method}
\usage{
\method{bread}{tsarma.estimate}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsarma.estimate}.}

\item{...}{not currently used.}
}
\value{
The analytic hessian of the model.
}
\description{
Bread Method
}
\author{
Alexios Galanos
}
