% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.tsarma.estimate}
\alias{coef.tsarma.estimate}
\alias{coef}
\title{Extract Model Coefficients}
\usage{
\method{coef}{tsarma.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsarma.estimate}.}

\item{...}{not currently used.}
}
\value{
A numeric named vector of estimated coefficients.
}
\description{
Extract the estimated coefficients of a model.
}
