% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{estimate.tsarma.spec}
\alias{estimate.tsarma.spec}
\alias{estimate}
\title{Estimates an ARMA model given a specification object using maximum likelihood and autodiff}
\usage{
\method{estimate}{tsarma.spec}(object, solver = "nloptr", control = NULL, ...)
}
\arguments{
\item{object}{an object of class tsarma.spec.}

\item{solver}{only \dQuote{nloptr} is currently supported (see \code{\link[nloptr]{nloptr}}).}

\item{control}{solver control parameters.}

\item{...}{not currently used.}
}
\value{
An object of class \dQuote{tsarma.estimate}.
}
\description{
Estimates an ARMA model given a specification object using maximum likelihood and autodiff
}
\details{
The underlying code is written using the TMB framework which uses
automatic differentiation and hence allows the generation of analytic
derivatives.
Stationarity of the process is controlled through a set of inequality constraints
on the characteristic roots of the AR/MA
The estimation makes 2 passes to the solver. The first pass uses no parameter
scaling, whilst in the second pass the parameters (as well as bounds) are scaled
making use of the estimated hessian from the first pass in order to generate
a hopefully more robust solution.
}
\author{
Alexios Galanos
}
