% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodel.R
\name{glance.tsarma.estimate}
\alias{glance.tsarma.estimate}
\alias{glance}
\title{Construct a summary glance of an estimated ARMA object}
\usage{
\method{glance}{tsarma.estimate}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsarma.estimate}.}

\item{...}{not currently used.}
}
\value{
A tibble().
}
\description{
glance method for tsarma
}
