% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{halflife.tsarma.estimate}
\alias{halflife.tsarma.estimate}
\alias{halflife}
\title{Half Life}
\usage{
\method{halflife}{tsarma.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsarma.estimate}.}

\item{...}{not currently used.}
}
\description{
Calculates and returns the half-life of an ARMA model.
}
\details{
The half life is defined as the period it
takes a series  to reach half its long-term average values. While this is know
for an AR(1) model, it is less clear what this should be for higher orders of
ARMA(p,q) models. Therefore, the function first converts the the ARMA model into
an infinite (truncated) MA model and then uses a spline approximation to find
the value for which the psi-weights (MA) equal 0.5.
}
