% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.tsarma.estimate}
\alias{logLik.tsarma.estimate}
\alias{logLik}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{tsarma.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsarma.estimate}.}

\item{...}{not currently used.}
}
\value{
An object of class \dQuote{logLik} with attributes for \dQuote{nobs} and
\dQuote{df}. The latter is equal to the number of estimated parameters
plus sum(ARMA order) (the initialization values).
}
\description{
Extract the log likelihood of the model at the estimated optimal
parameter values.
}
