% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.tsarma.estimate}
\alias{nobs.tsarma.estimate}
\alias{nobs}
\title{Extract the number of observations from an estimated model}
\usage{
\method{nobs}{tsarma.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsarma.estimate}.}

\item{...}{not currently used.}
}
\value{
A numeric value.
}
\description{
Extract the number of ‘observations’ from an estimated model.
This is principally intended to be used in computing BIC and used in other
tidy methods
}
