% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pit.tsarma.estimate}
\alias{pit.tsarma.estimate}
\alias{pit}
\title{Probability Integral Transform (PIT)}
\usage{
\method{pit}{tsarma.estimate}(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
An xts vector of the conditional probabilities.
}
\description{
Calculates and returns the conditional probability integral
transform given the data and estimated density
}
