% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.arma.irf}
\alias{plot.arma.irf}
\title{ARMA Impulse Response Function Plot}
\usage{
\method{plot}{arma.irf}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \dQuote{arma.irf}.}

\item{y}{not used.}

\item{...}{not used.}
}
\description{
Plot method for \dQuote{arma.irf} class.
}
