% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.arma.roots}
\alias{plot.arma.roots}
\title{ARMA Inverse Roots Plot}
\usage{
\method{plot}{arma.roots}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \dQuote{arma.roots}.}

\item{y}{not used.}

\item{...}{not used.}
}
\description{
Plot method for \dQuote{arma.roots} class.
}
