% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tsarma.estimate}
\alias{plot.tsarma.estimate}
\title{Estimated Model Plots}
\usage{
\method{plot}{tsarma.estimate}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsarma.estimate}.}

\item{y}{not used.}

\item{...}{not used.}
}
\description{
Plot method for \dQuote{tsarma.estimate} class.
}
