% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.tsarma.estimate}
\alias{predict.tsarma.estimate}
\alias{predict}
\title{Model Prediction}
\usage{
\method{predict}{tsarma.estimate}(
  object,
  h = 1,
  newxreg = NULL,
  bootstrap = FALSE,
  nsim = 1000,
  innov = NULL,
  innov_type = "r",
  innov_init = NULL,
  forc_dates = NULL,
  series_init = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsarma.estimate}.}

\item{h}{the forecast horizon.}

\item{newxreg}{regressors rows equal to h. This can be either
a numeric or xts matrix. Only needed if the model was estimated with regressors in the
conditional mean equation.}

\item{bootstrap}{whether to generate a bootstrap distribution for the conditional
volatility using re-sampling of the empirical innovations as in the paper by Pascual et al (2006).}

\item{nsim}{the number of simulations to use for generating the simulated
predictive distribution.}

\item{innov}{an optional matrix of dimensions nsim by h of innovations (see
innov_type below for types supported).}

\item{innov_type}{there are 3 options available. If type is \dQuote{q}, then
these represent quantiles and will be transformed to the appropriate distribution
innovations used in the model. If type is \dQuote{z}, these represent standardized
innovations which will be scaled by the model standard deviation. Finally, if
type is \dQuote{r}, then these are zero mean non-scaled innovations for which
the sigma has already been infused into the distribution (i.e. either static
or from a GARCH model output).}

\item{innov_init}{an optional vector of initialization values for the
standardized innovations. This allows the seeding of the initial innovations
with user supplied values (useful when simulating forward from an existing
model for the purpose of continuing the modeled series from some fixed point).
These should be of the same type as defined by innov_type.}

\item{forc_dates}{an optional vector of forecast dates equal to h. If NULL will use the
implied periodicity of the data to generate a regular sequence of dates after the
last available date in the data.}

\item{series_init}{an optional vector of values (y) to initialize the forecast.
If NULL, will use the last available values from the model. This must
be equal to the max of the ARMA order.}

\item{...}{additional arguments for future expansion options.}
}
\value{
A \dQuote{tsarma.predict} object.
}
\description{
Prediction function for class \dQuote{tsarma.estimate}.
}
\details{
The bootstrap method considered here, is based on re-sampling innovations
from the empirical distribution of the fitted ARMA model to generate future
realizations of the series.
}
