% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sigma.tsarma.estimate}
\alias{sigma.tsarma.estimate}
\alias{sigma}
\title{Extract Model Standard Deviation}
\usage{
\method{sigma}{tsarma.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsarma.estimate}.}

\item{...}{not currently used.}
}
\value{
An numeric vector of length 1.
}
\description{
Extract the standard deviation from a ARMA model.
}
