% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simulate.tsarma.spec}
\alias{simulate.tsarma.spec}
\alias{simulate}
\alias{simulate.tsarma.estimate}
\title{Model Simulation}
\usage{
\method{simulate}{tsarma.spec}(
  object,
  nsim = 1000,
  seed = NULL,
  h = 1,
  innov = NULL,
  innov_init = NULL,
  innov_type = "r",
  series_init = NULL,
  xreg = NULL,
  ...
)

\method{simulate}{tsarma.estimate}(
  object,
  nsim = 1000,
  seed = NULL,
  h = 1,
  innov = NULL,
  innov_init = NULL,
  series_init = NULL,
  xreg = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsarma.spec} or \dQuote{tsarma.estimate}.}

\item{nsim}{the number of sample paths to generate.}

\item{seed}{an integer that will be used in a call to set.seed before simulating.}

\item{h}{the number of time steps to simulate paths for.}

\item{innov}{an optional matrix of dimensions nsim by h of innovations (see
innov_type below for types supported).}

\item{innov_init}{an optional vector of initialization values for the
standardized innovations. This allows the seeding of the initial innovations
with user supplied values (useful when simulating forward from an existing
model for the purpose of continuing the modeled series from some fixed point).
These should be of the same type as defined by innov_type.}

\item{innov_type}{there are 3 options available. If type is \dQuote{q}, then
these represent quantiles and will be transformed to the appropriate distribution
innovations used in the model. If type is \dQuote{z}, these represent standardized
innovations which will be scaled by the model standard deviation. Finally, if
type is \dQuote{r}, then these are zero mean non-scaled innovations for which
the sigma has already been infused into the distribution (i.e. either static
or from a GARCH model output).}

\item{series_init}{the seed value for initializing the mean equation recursion.
If NULL, the unconditional mean is used based on the supplied parameters.
This should be a vector and assumes all sample paths are seeded the same way.}

\item{xreg}{an optional vector of length h representing any mean regressors
to use in the simulation.}

\item{...}{not currently used.}
}
\value{
An object of class \dQuote{tsarma.simulate} with slots for the
simulated series and innovations, both of which are of class
\dQuote{tsmodel.distribution}.
}
\description{
Simulates paths of a ARMA model.
}
