% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsarma-package.R
\docType{package}
\name{tsarma-package}
\alias{tsarma}
\alias{tsarma-package}
\title{tsarma: Univariate Autoregressive Moving Average ('ARMA') Models}
\description{
Autoregressive Moving Average Models with methods for estimation using automatic differentiation, filtering, prediction, simulation and diagnostics.
}
\keyword{internal}
