% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsfilter.tsarma.estimate}
\alias{tsfilter.tsarma.estimate}
\alias{tsfilter}
\alias{tsfilter.tsarma.spec}
\title{Model Filtering}
\usage{
\method{tsfilter}{tsarma.estimate}(object, y = NULL, newxreg = NULL, ...)

\method{tsfilter}{tsarma.spec}(object, y = NULL, newxreg = NULL, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsarma.estimate} or \dQuote{tsarna.spec}.}

\item{y}{an xts vector of new values to filter.}

\item{newxreg}{model regressors with the same number of rows as y. This can be either
a numeric or xts matrix. Only needed if the model was estimated with regressors in the
conditional mean equation.}

\item{...}{additional arguments for future expansion options.}
}
\value{
A \dQuote{tsarma.estimate} object with updated information if the input
object was also of class \dQuote{tsarma.estimate}, else will return an object
of class \dQuote{tsarma.filter} for input of class \dQuote{tsarma.spec}.
}
\description{
Filters new data based on an already estimated model.
}
\details{
The method filters new data and updates the object with this new information,
appending y and xreg datasets so that is can be called recursively as new
data arrives. For a \dQuote{tsarma.spec} object as input, the existing data
will be filtered with the parameter in the specification (considered fixed), but
the returned object will not contain information about parameter uncertainty
such as the hessian or scores.
}
