% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{unconditional.tsarma.estimate}
\alias{unconditional.tsarma.estimate}
\alias{unconditional}
\title{Unconditional Value}
\usage{
\method{unconditional}{tsarma.estimate}(object, ...)
}
\arguments{
\item{object}{an object.}

\item{...}{additional parameters passed to the method.}
}
\value{
A scalar of the unconditional value. For ARMA models this represents
the unconditional mean.
}
\description{
General method the unconditional value of a model.
}
