% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsarma-package.R
\docType{package}
\name{tsarma-package}
\alias{tsarma}
\alias{tsarma-package}
\title{tsarma: Univariate ARMA Models}
\description{
Autoregressive Moving Average Models. Estimation, filtering, prediction, simulation and diagnostics.
}
\keyword{internal}
